chore: move functions to route groups
