+++ ui/src/app/(auth)/initialize/page.tsx
@@ -1,13 +1,13 @@
 "use client"
 
 import { useState, ChangeEvent } from "react"
-import { getStatus, login, postFirstUser } from "../queries"
+import { getStatus, login, postFirstUser } from "../../queries"
 import { useMutation, useQuery } from "@tanstack/react-query"
 import { useRouter } from "next/navigation"
-import { passwordIsValid } from "../utils"
-import { useAuth } from "../auth/authContext"
+import { passwordIsValid } from "../../utils"
+import { useAuth } from "../../auth/authContext"
 import { useCookies } from "react-cookie"
-import { statusResponse } from "../types"
+import { statusResponse } from "../../types"
 import { Input, PasswordToggle, Button, Form, LoginPageLayout } from "@canonical/react-components";
