+++ ui/src/app/(auth)/login/page.tsx
@@ -1,12 +1,12 @@
 "use client"
 
-import { getStatus, login } from "../queries"
+import { getStatus, login } from "../../queries"
 import { useMutation, useQuery } from "@tanstack/react-query"
 import { useState, ChangeEvent } from "react"
 import { useCookies } from "react-cookie"
 import { useRouter } from "next/navigation"
-import { useAuth } from "../auth/authContext"
-import { statusResponse } from "../types"
+import { useAuth } from "../../auth/authContext"
+import { statusResponse } from "../../types"
 import { Input, PasswordToggle, Button, Form, Notification, LoginPageLayout } from "@canonical/react-components";
