+++ ui/src/app/(notary)/certificate_requests/page.tsx
@@ -1,23 +1,24 @@
 "use client"
 
 import { useQuery } from "@tanstack/react-query"
-import { ListUsers } from "../queries"
-import { UserEntry } from "../types"
+import { CertificateRequestsTable } from "./table"
+import { getCertificateRequests } from "../../queries"
+import { CSREntry } from "../../types"
 import { useCookies } from "react-cookie"
 import { useRouter } from "next/navigation"
-import { UsersTable } from "./table"
-import Loading from "../loading"
-import Error from "../error"
+import Loading from "../../loading"
+import Error from "../../error"
 
-export default function Users() {
+
+export default function CertificateRequests() {
     const router = useRouter()
     const [cookies, setCookie, removeCookie] = useCookies(['user_token']);
     if (!cookies.user_token) {
         router.push("/login")
     }
-    const query = useQuery<UserEntry[], Error>({
-        queryKey: ['users', cookies.user_token],
-        queryFn: () => ListUsers({ authToken: cookies.user_token }),
+    const query = useQuery<CSREntry[], Error>({
+        queryKey: ['csrs', cookies.user_token],
+        queryFn: () => getCertificateRequests({ authToken: cookies.user_token }),
         retry: (failureCount, error): boolean => {
             if (error.message.includes("401")) {
                 return false
@@ -32,6 +33,8 @@ export default function Users() {
         }
         return <Error msg={query.error.message} />
     }
-    const users = Array.from(query.data ? query.data : [])
-    return <UsersTable users={users} />
+    const csrs = Array.from(query.data ? query.data : [])
+    return (
+        <CertificateRequestsTable csrs={csrs} />
+    )
 }
