+++ ui/src/app/(notary)/users/asideForm.tsx
@@ -1,9 +1,9 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { passwordIsValid } from "../utils";
-import { changePassword, postUser } from "../queries";
+import { passwordIsValid } from "../../utils";
+import { changePassword, postUser } from "../../queries";
 import { ChangeEvent, useContext, useState } from "react";
-import { AsideContext } from "../aside";
-import { useAuth } from "../auth/authContext";
+import { AsideContext } from "../../aside";
+import { useAuth } from "../../auth/authContext";
 import { Panel, Button, Input, PasswordToggle, Form } from "@canonical/react-components";
 
 export default function UsersPageAsidePanel() {
