+++ ui/src/app/(notary)/users/components.tsx
@@ -1,8 +1,8 @@
 import { Dispatch, SetStateAction, useState, ChangeEvent, createContext } from "react"
-import { useAuth } from "../auth/authContext"
+import { useAuth } from "../../auth/authContext"
 import { useMutation, useQueryClient } from "@tanstack/react-query"
-import { changePassword } from "../queries"
-import { passwordIsValid } from "../utils"
+import { changePassword } from "../../queries"
+import { passwordIsValid } from "../../utils"
 import { ConfirmationModal, Modal, Button, Input, PasswordToggle, Form } from "@canonical/react-components";
 
 export type ConfirmationModalData = {
